fix interceptor
This is a working implementation for REST and GQL
using `ModuleRef` instead of relying on REQUEST
scoped interceptors. You need to make sure that
the gql context has req added to it (it does by
default for apollo-server-express, but being
explicit never hurt either). You can add that in
by doing

```ts
GraphQLModule.forRoot({
  context: ({ req }) => ({ req })
  // other options
})
```

This will allow for `gql.getContext().req`
to pull back the request object which is require
for `ContextIdFactory.getByRequest()`.

ref: [#5722](https://github.com/nestjs/nest/issues/5722)